refactor: extract middleware and better-auth factories
