+++ apps/auth/src/better-auth/server.ts
